requirements.txt
# HTTP client
httpx>=0.27.0
tenacity>=8.2.0

# Config
pyyaml>=6.0

# Data
pandas>=2.2.0
numpy>=1.26.0
pyarrow>=15.0.0

# ML — core
scikit-learn>=1.4.0
xgboost>=2.0.0
shap>=0.45.0
optuna>=3.6.0

# ML — survival analysis
lifelines>=0.29.0

# ML — optimization
cvxpy>=1.4.0

# Serving
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0

# NullClaw
anthropic>=0.28.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
httpx>=0.27.0  # also used by TestClient
